renamed main page to get-started
