+++ apps/web/src/components/GoToCommunity/StartBuilding.tsx
@@ -3,7 +3,7 @@ import {
   sectionContainer,
   titleClasses,
   START_BUILDING_SECTION_ID,
-} from '../../../app/(base-org)/go-to-community/page';
+} from '../../../app/(base-org)/get-started/page';
 import { ButtonWithLinkAndEventLogging } from '../Button/ButtonWithLinkAndEventLogging';
 import ResourceGrid from './ResourceGrid/ResourceGrid';
 import { ResourceSectionType } from './resourceTypes';
